Rename `command_line` to `shell`
Several files were updated to change the `command_line` namespace and directory to `shell`. Besides renaming files and directories, the changes involved also updating references within the files, like adjusting the namespace and includes. There was also a slight code update in `terminal.cpp`.
